app/requirements.txt
banana-dev==4.0.2
streamlit==1.22.0
streamlit-drawable-canvas==0.9.2
PyYAML==5.4.1
